[presentation-rooms] poped speech recognition into a new component in order to made it more modular
